물품 예외처리 초안

+++ backend/src/main/java/com/modimoa/backend/service/UserService.java
@@ -27,38 +27,39 @@ public class UserService {
         String result = "";
 
         Optional <User> user = userRepository.findByUserEmail(userEmail);
+
+
          if(user.isPresent()){
+
+             //예외처리(409)
             result = userEmail+"은 이미 존재합니다.";
         }else{
             user.orElseGet(() ->userRepository.save(new User(userEmail,userImage,oauthCookie,"new","new")));
-            result = userEmail+"로 회원가입 되었습니다.";
+            result = "success";
         }
         return  result;
     }
 
 
-    public String[] login(String token) throws NoSuchAlgorithmException {
+    public String login(String userEmail) throws NoSuchAlgorithmException {
         String result ="";
 
-        Optional <User> user = Optional.of(userRepository.findByOauthToken(token).get());
+        Optional <User> user = Optional.of(userRepository.findByUserEmail(userEmail).get());
 
         String userId=user.get().getUserEmail();
 
-        String accessCrypt = EncryptionUtils.encryptSHA256(token+userId+"access");
-        String refreshCrypt= EncryptionUtils.encryptSHA256(token+userId+"refresh");
-
-
-        String[] tokenAry ={accessCrypt,refreshCrypt};
+        String accessToken = EncryptionUtils.encryptSHA256(userEmail+userId+"access");
+        String refreshToken= EncryptionUtils.encryptSHA256(userEmail+userId+"refresh");
 
         if(user.isPresent()){
-
-            user.get().updateTokens(accessCrypt,refreshCrypt);
-            result = user.get().getUserEmail()+"로 로그인 되었습니다."+" access: "+accessCrypt+" refresh: "+refreshCrypt;
+            user.get().updateTokens(accessToken,refreshToken);
+            result = user.get().getUserEmail()+"로 로그인 되었습니다."+" access: "+accessToken+" refresh: "+refreshToken;
         }else{
+            //예외처리(유저 존재하지 않음403 forbidden)
             result = "로그인에 실패하셨습니다.";
         }
 
-        return tokenAry;
+        return accessToken;
     }
 
 
@@ -68,7 +69,7 @@ public class UserService {
     };
 
 
-    public String logout(String token) {
+    public void logout(String token) {
         String result ="";
 
         Optional <User> user = Optional.of(userRepository.findByOauthToken(token).get());
@@ -77,21 +78,23 @@ public class UserService {
             String accessToken="";
             String refreshToken="";
             user.get().updateTokens(accessToken,refreshToken);
-            result = user.get().getUserEmail()+"이 로그아웃 되었습니다.";
+
         }else{
-            result = "로그아웃에 실패하셨습니다.";
+            //예외처리(403 forbidden)
         }
 
-        return result;
     }
 
-    public String withdrawal(String token) {
+    public void withdrawal(String token) {
 
-        User user = userRepository.findByAccessToken(token).get();
+        Optional <User> user = Optional.of(userRepository.findByAccessToken(token).get());
 
-        userRepository.deleteByAccessToken(token);
+        if(user.isPresent()){
+            userRepository.deleteByAccessToken(token);
 
-        return "탈퇴에 성공했습니다.";
+        }else{
+            //예외처리 403forbidden
+        }
 
     }
 
@@ -99,12 +102,15 @@ public class UserService {
     public Map getUserInfo(String token) {
         Map result= new HashMap<String,String>();
 
-        User user = userRepository.findByAccessToken(token).get();
+        Optional <User> user = Optional.of(userRepository.findByAccessToken(token).get());
 
-        result.put("user_email", user.getUserEmail());
-        result.put("user_image", user.getUserImage());
+        if(user.isPresent()){
+            result.put("user_email", user.get().getUserEmail());
+            result.put("user_image", user.get().getUserImage());
 
-
+        }else{
+            //예외처리 403forbidden
+        }
         return result;
 
     }
